\name{SMLoutliers-package}
\alias{SMLoutliers-package}
\alias{SMLoutliers}
\docType{package}
\title{
An R package for identifying outliers using statistical and machine learning methods
}
\description{
We intend to provide host of methods for identifying outliers. This will cut across statistical and machine learning methods. 
}
\references{
M.M. Breunig, H.P. Kriegel, R.T. Ng, and J. Sander. Lof: Identifying density-based
local outliers. In Proc. SIGMOD Conf., pages 93-104, 2000.
}
\keyword{outliers}
\keyword{LOCI}

\examples{
data(stiff)
summary(stiff)
}
